backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.6
pytz==2023.3
pyyaml==6.0.1